OPENSTACK-358 Pluggable IPAM blueprint code
Currently most (if not all) Neutron plugins leverage an IPAM
implementation that is embedded in the db_plugin base implementation.
While this works well in a self contained system, it makes it difficult
to integrate with an external IPAM backend.

The proposal is to hash out an IPAM interface derived from the existing
IPAM-related interfaces defined in the db_plugin and refactor the
current implementation into a configurable IPAM driver. One can then
envision implementing and configuring additional independent drivers to
external IPAM backends as well.

Change-Id: Ia60873db9f9123dc5fa5cdeb7d714a910f3e1118